Refactor Taglines
