+++ Cargo.toml
@@ -74,6 +74,7 @@ members = [
   "crates/db_views/report_combined",
   "crates/db_views/search_combined",
   "crates/db_views/site",
+  "crates/db_views/tagline",
   "crates/routes",
   "crates/federate",
   "crates/email",
@@ -141,6 +142,7 @@ lemmy_db_views_report_combined = { version = "=1.0.0-alpha.5", path = "./crates/
 lemmy_db_views_reports = { version = "=1.0.0-alpha.5", path = "./crates/db_views/reports" }
 lemmy_db_views_search_combined = { version = "=1.0.0-alpha.5", path = "./crates/db_views/search_combined" }
 lemmy_db_views_site = { version = "=1.0.0-alpha.5", path = "./crates/db_views/site" }
+lemmy_db_views_tagline = { version = "=1.0.0-alpha.5", path = "./crates/db_views/tagline" }
 lemmy_db_views_vote = { version = "=1.0.0-alpha.5", path = "./crates/db_views/vote" }
 activitypub_federation = { version = "0.7.0-beta.1", default-features = false, features = [
   "actix-web",
