+++ crates/api/api_common/Cargo.toml
@@ -73,4 +73,5 @@ lemmy_db_views_report_combined.workspace = true
 lemmy_db_views_reports.workspace = true
 lemmy_db_views_search_combined.workspace = true
 lemmy_db_views_site.workspace = true
+lemmy_db_views_tagline.workspace = true
 lemmy_db_views_vote.workspace = true
