+++ crates/api/api_common/src/tagline.rs
@@ -1,6 +1,6 @@
 pub use lemmy_db_schema::{newtypes::TaglineId, source::tagline::Tagline};
-pub use lemmy_db_views_site::api::{ListTaglines, ListTaglinesResponse, TaglineResponse};
+pub use lemmy_db_views_tagline::{ListTaglines, ListTaglinesResponse, TaglineResponse};
 
 pub mod aministration {
-  pub use lemmy_db_views_site::api::{CreateSite, DeleteTagline, UpdateTagline};
+  pub use lemmy_db_views_tagline::{CreateTagline, DeleteTagline, UpdateTagline};
 }
