+++ crates/api/api_crud/src/tagline/delete.rs
@@ -4,7 +4,7 @@ use lemmy_api_utils::{context::LemmyContext, utils::is_admin};
 use lemmy_db_schema::{source::tagline::Tagline, traits::Crud};
 use lemmy_db_views_api_misc::SuccessResponse;
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_site::api::DeleteTagline;
+use lemmy_db_views_tagline::DeleteTagline;
 use lemmy_utils::error::LemmyError;
 
 pub async fn delete_tagline(
