+++ crates/db_views/tagline/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "lemmy_db_views_tagline"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+license.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[lib]
+doctest = false
+
+[lints]
+workspace = true
+
+[features]
+full = ["lemmy_db_schema/full"]
+ts-rs = ["dep:ts-rs"]
+
+[dependencies]
+lemmy_db_schema = { workspace = true }
+serde = { workspace = true }
+serde_with = { workspace = true }
+ts-rs = { workspace = true, optional = true }
